Move style files to the assets and change paths for the imports
