+++ src/assets/scss/index.scss
@@ -2,22 +2,22 @@
 
 @font-face {
   font-family: "Montserrat-Bold";
-  src: url(assets/fonts/Montserrat-Bold.ttf);
+  src: url(../fonts/Montserrat-Bold.ttf);
 }
 
 @font-face {
   font-family: "Montserrat-Regular";
-  src: url(assets/fonts/Montserrat-Regular.ttf);
+  src: url(../fonts/Montserrat-Regular.ttf);
 }
 
 @font-face {
   font-family: "RobotoSlabLight";
-  src: url(assets/fonts/RobotoSlab-Light.ttf);
+  src: url(../fonts/RobotoSlab-Light.ttf);
 }
 
 @font-face {
   font-family: "RobotoSlabBold";
-  src: url(assets/fonts/RobotoSlab-Bold.ttf);
+  src: url(../fonts/RobotoSlab-Bold.ttf);
 }
 
 body {
