+++ src/assets/scss/progressBar.scss
