+++ src/components/CategoryFilter.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import '../CategoryFilter.scss';
+import '../assets/scss/CategoryFilter.scss';
 import categoriesList from '../const/categories';
 
 const CategoryFilter = ({ onFilter }) => {
