+++ src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import './reset.css';
-import './index.scss';
+import './assets/css/reset.css';
+import './assets/scss/index.scss';
 import App from './components/App';
 import rootReducer from './reducers/index';
 import categoriesList from './const/categories';
